# go/app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple Mercari Items API.",
        "title": "Simple Mercari Items API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:1313",
    "basePath": "/",
    "paths": {
        "/image/:itemImg": {
            "get": {
                "description": "find an item by id",
                "produces": [
                    "application/json"
                ],
                "summary": "get an image",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Item's image name",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/items": {
            "get": {
                "description": "get all items",
                "produces": [
                    "application/json"
                ],
                "summary": "get items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Items"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "description": "post an item to its table.",
                "produces": [
                    "application/json"
                ],
                "summary": "post an item",
                "parameters": [
                    {
                        "description": "Item's name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Item's category",
                        "name": "category",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/items/:id": {
            "get": {
                "description": "find an item by id",
                "produces": [
                    "application/json"
                ],
                "summary": "find an item",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item's id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "obejct"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/items/search": {
            "get": {
                "description": "search Items by name",
                "produces": [
                    "application/json"
                ],
                "summary": "search Items by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Keyword to match Item's name",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Items"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Item": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "main.Items": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.Item"
                    }
                }
            }
        },
        "main.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}